requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
attrs==25.4.0
beautifulsoup4==4.14.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
curl_cffi==0.14.0
frozendict==2.4.7
frozenlist==1.8.0
greenlet==3.2.4
h11==0.16.0
idna==3.11
lxml==6.0.2
multidict==6.7.0
multitasking==0.0.12
numpy==2.3.5
outcome==1.3.0.post0
pandas==2.3.3
peewee==3.18.3
platformdirs==4.5.1
playwright==1.56.0
playwright-stealth==2.0.0
propcache==0.4.1
protobuf==6.33.2
psycopg2-binary==2.9.11
pycparser==2.23
pyee==13.0.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
selenium==4.38.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
SQLAlchemy==2.0.44
tabulate==0.9.0
tqdm==4.67.1
trio==0.32.0
trio-websocket==0.12.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
websocket-client==1.9.0
websockets==15.0.1
wsproto==1.3.2
yarl==1.22.0
yfinance==0.2.66
